requirements.txt
Django~=4.0.6
Pillow~=9.3.0
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7

djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1

django-debug-toolbar==3.7.0

mysql-client==0.0.1
mysqlclient==2.1.1
PyMySQL==1.0.2
sqlparse==0.4.2


